fixed bug with url query not showing in input box
